Splitting and moving Setup classes
